Move ui folders to an UI folder
