+++ src/book/pages/BooksPage/BooksPage.tsx
@@ -4,8 +4,8 @@ import { useSearchParams } from "react-router";
 import useBooks from "../../hooks/useBooks";
 import useLoading from "../../../hooks/useLoading";
 import Bookshelf from "../../components/Bookshelf/Bookshelf";
-import Pagination from "../../../components/Pagination/Pagination";
-import Loading from "../../../components/Loading/Loading";
+import Pagination from "../../../ui/components/Pagination/Pagination";
+import Loading from "../../../ui/components/Loading/Loading";
 import BookFilter from "../../components/BookFilter/BookFilter";
 import type { BookFilters } from "../../types";
 import "./BooksPage.css";
