+++ src/ui/components/Layout/Layout.test.tsx
@@ -1,13 +1,16 @@
 import { render, screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import ContextProvider from "../../test-utils/ContextProvider";
+import ContextProvider from "../../../test-utils/ContextProvider";
 import { http, HttpResponse } from "msw";
-import { server } from "../../book/mocks/node";
-import AppTestRouter from "../../router/AppTestRouter";
+import { server } from "../../../book/mocks/node";
+import AppTestRouter from "../../../router/AppTestRouter";
 import Layout from "./Layout";
-import type { BooksInfoDto } from "../../book/client/types";
-import { demonSlayerVol1, onePieceVol1 } from "../../book/fixtures/fixturesDto";
-import { dragonBallRead } from "../../book/fixtures/fixtures";
+import type { BooksInfoDto } from "../../../book/client/types";
+import {
+  demonSlayerVol1,
+  onePieceVol1,
+} from "../../../book/fixtures/fixturesDto";
+import { dragonBallRead } from "../../../book/fixtures/fixtures";
 
 const user = userEvent.setup();
 window.scrollTo = vitest.fn();
