+++ src/ui/components/Layout/Layout.tsx
@@ -3,7 +3,7 @@ import { Outlet } from "react-router";
 import Modal from "../Modal/Modal";
 import Header from "../Header/Header";
 import Navigation from "../Navigation/Navigation";
-import useModal from "../../hooks/useModal";
+import useModal from "../../../hooks/useModal";
 import "./Layout.css";
 
 const Layout: React.FC = () => {
